keypoints_regression/aws/src/requirements.txt
ultralytics
pyaml
numpy==1.26.4
